requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cloudinary==1.29.0
distlib==0.3.4
Django==4.0.3
djangorestframework==3.13.1
filelock==3.6.0
nine==1.1.0
Pillow==9.1.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.0
